pyproject.toml
[project]
name = "sms-dashboard"
license = "MIT"
version = "0.1.0"
description = "A web-based dashboard for managing and monitoring SMS messages, built with Flask and MySQL."
authors = [
    { name = "Mohammad Ahmadi", email = "mohammad.ahmadi98@proton.me" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Flask>=2.0",
    "mysql-connector-python>=8.0",
    "python-dotenv>=1.1.1",
    "gunicorn>=23.0.0",
    "python-telegram-bot (>=22.4,<23.0)"
]

[tool.poetry.scripts]
sms-dash = "app:main"
sms-dev = "sms-dashboard.run_dev:main"
sms-prod = "sms-dashboard.run_production:main"



[tool.poetry]
packages = [{ include = "sms-dashboard", from = "src" }]
package-mode = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"